requirements.txt
fastapi
uvicorn
gunicorn
google-generativeai
Pillow
PyMuPDF
python-multipart